+++ test_new_module.py
@@ -1,1 +1,3 @@
-
+import src.budget_manager
+import src.database
+import src.main
