+++ tests/test_new_module.py
@@ -1,0 +1,1 @@
+
